fix: galaxy number_of_stars INT -> BIGINT (1e12 overflow) 2026-08-31

+++ 01-celestial-bodies/dump.sql
@@ -25,7 +25,7 @@ CREATE TABLE galaxy (
   description TEXT,
   type VARCHAR(50),
   distance_from_earth NUMERIC,
-  number_of_stars INT
+  number_of_stars BIGINT
 );
 
 -- star (FK -> galaxy)
